Cambios Endpoints
Se añaden los endpoint de los himnos.
Se cambia la lógica de listado. 
Se añade campo letra y titulo a los himnos. 
Se añade campo descripción a los ambientes

+++ api150/api/pray/delete.php
@@ -21,23 +21,36 @@
     $data = json_decode(file_get_contents("php://input"));
 
     $idRecibida = htmlspecialchars($_GET["idOracion"]);
-    
-    if (!empty($idRecibida)) {
-        // Tengo todos los datos
-        // Comprobamos que la id corresponde a un registro 
-        if ($cf->comprobarExisteOracionPorId($idRecibida)) {
-            // Efectivamente existe y se puede eliminar
-            $query = "DELETE FROM oraciones WHERE id_Oracion like ".$idRecibida.";";
-            $stmt = $database->getConn()->prepare($query);
-            $stmt->execute();
-            echo json_encode(array(" error : 0, message : Elemento eliminado"));
+
+    $token = htmlspecialchars($_GET["token"]);
+
+    // Comprobamos que tiene permisos de administrador
+    if ($cf->comprobarTokenAdmin($token) == 1) { 
+        if (!empty($idRecibida)) {
+            // Tengo todos los datos
+            // Comprobamos que la id corresponde a un registro 
+            if ($cf->comprobarExisteOracionPorId($idRecibida)) {
+                // Efectivamente existe y se puede eliminar
+                $query = "DELETE FROM oraciones WHERE id_Oracion like ".$idRecibida.";";
+                $stmt = $database->getConn()->prepare($query);
+                $stmt->execute();
+                echo json_encode(array(" error : 0, message : Elemento eliminado"));
+            } else {
+                //No existe y por lo tanto no se puede eliminar 
+                echo json_encode(array("error : 1, message : El registro no existe"));
+            }
         } else {
-            //No existe y por lo tanto no se puede eliminar 
-            echo json_encode(array("error : 1, message : El registro no existe"));
+            echo json_encode("error : 1, message : faltan uno o más datos");
         }
+
+
+    } elseif ($cf->comprobarTokenAdmin($token) == 0) {
+        echo json_encode("error : 2, message : no tiene permisos para realizar esta operación");
     } else {
-        echo json_encode("error : 1, message : faltan uno o más datos");
+        echo json_encode("error : 3, message : token no valido");
     }
+    
+
